# swagger/apiserver/go/routers.go
/*
 * Chain Query
 *
 * The LBRY blockchain is read into SQL where important structured information can be extracted through the Chain Query API.
 *
 * API version: 0.1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

import (
	"net/http"
	"strings"

	. "github.com/irmf/chainquery/apiactions"
	. "github.com/lbryio/lbry.go/extras/api"

	"github.com/gorilla/mux"
	"github.com/prometheus/client_golang/prometheus/promhttp"
)

type Route struct {
	Name        string
	Method      string
	Pattern     string
	HandlerFunc Handler
}

type Routes []Route

func NewRouter() *mux.Router {
	router := mux.NewRouter().StrictSlash(true)
	for _, route := range routes {
		var handler http.Handler
		handler = route.HandlerFunc
		handler = Logger(handler, route.Name)

		router.
			Methods(route.Method).
			Path(route.Pattern).
			Name(route.Name).
			Handler(handler)
	}

	router.Handle("/metrics", promBasicAuthWrapper(promhttp.Handler()))

	return router
}

var routes = Routes{
	Route{
		"Index",
		"GET",
		"/api/",
		IndexAction,
	},

	Route{
		"AutoUpdate",
		strings.ToUpper("Post"),
		"/api/autoupdate",
		AutoUpdateAction,
	},

	Route{
		"SQLQuery",
		strings.ToUpper("Get"),
		"/api/sql",
		SQLQueryAction,
	},

	Route{
		"AddressSummary",
		strings.ToUpper("Get"),
		"/api/addresssummary",
		AddressSummaryAction,
	},

	Route{
		"ChainQueryStatus",
		strings.ToUpper("Get"),
		"/api/status",
		ChainQueryStatusAction,
	},

	Route{
		"ValidateChain",
		strings.ToUpper("Get"),
		"/api/validate",
		ValidateChainData,
	},

	Route{
		"ProcessBlocks",
		strings.ToUpper("Get"),
		"/api/process",
		ProcessBlocks,
	},

	Route{
		"SyncAddressBalance",
		strings.ToUpper("Get"),
		"/api/sync/addresses",
		SyncAddressBalance,
	},

	Route{
		"SyncClaimName",
		strings.ToUpper("Get"),
		"/api/sync/name",
		SyncName,
	},

	Route{
		"SyncTransactionValue",
		strings.ToUpper("Get"),
		"/api/sync/txvalues",
		SyncTransactionValue,
	},
}

var PromPassword string
var PromUser string

func promBasicAuthWrapper(h http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		user, pass, ok := r.BasicAuth()
		if !ok {
			http.Error(w, "authentication required", http.StatusBadRequest)
			return
		}
		if PromUser == "" {
			http.Error(w, "not configured", http.StatusPreconditionFailed)
		} else {
			if user == PromUser && pass == PromPassword {
				h.ServeHTTP(w, r)
			} else {
				http.Error(w, "invalid username or password", http.StatusForbidden)
			}
		}
	})
}